// extensions/vuengine-studio-extension/src/editors/browser/components/Common/Dotting/components/Canvas/types.ts
export interface Coord {
  x: number;
  y: number;
}

export interface Dimensions {
  width: number;
  height: number;
}

export type ButtonDimensions = Coord & Dimensions;

export interface PanZoom {
  scale: number;
  offset: Coord;
}

export interface PixelData {
  color: string;
}

export interface ImageDownloadOptions {
  isGridVisible?: boolean;
  type: 'png' | 'svg';
}

export type DottingData = Map<number, Map<number, PixelData>>;